Preparing client to support multiple versions
