+++ ecsclient/common/exceptions.py
