+++ ecsclient/common/util.py
