+++ ecsclient/v2/configuration/certificate.py
