+++ ecsclient/v2/configuration/configuration_properties.py
