+++ ecsclient/v2/user_management/secret_key_self_service.py
